+++ crates/cxx-qt-lib/src/core/qhash/mod.rs
@@ -206,6 +206,19 @@ where
     }
 }
 
+impl<'a, T> IntoIterator for &'a QHash<T>
+where
+    T: QHashPair,
+{
+    type Item = (&'a T::Key, &'a T::Value);
+
+    type IntoIter = Iter<'a, T>;
+
+    fn into_iter(self) -> Self::IntoIter {
+        self.iter()
+    }
+}
+
 /// Trait implementation for a pair in a [`QHash`].
 pub trait QHashPair: Sized {
     type Key;